fix: rename classes
